NaN check

+++ app.js
@@ -103,6 +103,10 @@ app.delete('/restaurants', function(req, res) {
 
 app.delete('/restaurants/:uid', function(req, res) {
     let requestID = req.params.uid;
+    if (isNaN(requestID)) {
+      throw ("NaN!");
+
+    }
     return Restaurant.destroy({
         where: {id: requestID}
       })
